testing/web-platform/tests/websockets/interfaces/WebSocket/bufferedAmount/004.html

<!doctype html>
<title>WebSockets: delete bufferedAmount</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
test(function(){
  delete WebSocket.prototype.bufferedAmount;
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo');
  assert_equals(ws.bufferedAmount, 0, 'after attempt to delete WebSocket.prototype.bufferedAmount');
  ws.close();
  delete ws.bufferedAmount;
  assert_equals(ws.bufferedAmount, 0, 'after attempt to delete ws.bufferedAmount');
});
</script>
